% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_analyses.R
\name{add_analyses}
\alias{add_analyses}
\title{Add Analyses To A Design Frame}
\usage{
add_analyses(
  obj,
  cluster = NA,
  engine = c("mplus", "lavaan", "lm"),
  estimator = c("wlsmv", "mlr", NA),
  ignore_x_type = FALSE,
  ignore_dummy01 = FALSE,
  warn_low_n = FALSE,
  remove_text_var = TRUE,
  remove_empty_y_var = TRUE,
  drop_duplicates = TRUE,
  path = tempdir()
)
}
\arguments{
\item{obj}{A list as returned from create_infoframe()}

\item{cluster}{Character string identifying the variable used as cluster-identifier.}

\item{engine}{either "mplus", "lavaan" or "lm". Currently only supports mplus.}

\item{estimator}{"Either "wlsmv", "mlr" or NA.}

\item{ignore_x_type}{Logical, default is FALSE. Whether to ignore the type 
of the independent variable (such as ordinal, nominal) and treat them 
as they are stored (e.g. model as continuous/binary).}

\item{ignore_dummy01}{Logical, default is FALSE. If FALSE, will automatically
adjust binary variables such that the lowest category is zero, and the 
highest is 1.}

\item{warn_low_n}{Logical, default is FALSE. If TRUE ????}

\item{remove_text_var}{Logical, default is TRUE. If TRUE, will automatically
drop all variables with type=="text", except for cluster-variable.}

\item{remove_empty_y_var}{Logical, default is TRUE. If TRUE, ???}

\item{drop_duplicates}{Logical, default is TRUE. If TRUE, will automatically
remove duplicate rows in design frame.}

\item{path}{Path to saving Mplus files. Default is a temporary folder.}
}
\value{
A data frame with model specifications in simplified form (y_var=Y, independent=>X), estimate of the regression coefficient if any, and model fit indices if any.
}
\description{
Given the specifications of design_frame in obj, run regression analyses.
}
