% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_barchart_mschart.R
\name{add_barchart_mschart}
\alias{add_barchart_mschart}
\title{Add Barchart}
\usage{
add_barchart_mschart(
  obj,
  prefix_number = FALSE,
  hide_label_if_less_than = 0,
  font_size = 8,
  round_digits = 1,
  show_n = FALSE,
  show_p = TRUE,
  sorting = NULL,
  drop_na_y = TRUE,
  drop_na_x = TRUE,
  type = 1L,
  drop_duplicates = TRUE,
  remove_empty_y_var = FALSE
)
}
\arguments{
\item{obj}{A list containing df, var_frame and design_frame, as created by
create_infoframe().}

\item{prefix_number}{Whether to add the underlying number index for the
categories as a prefix in the printed chart and table labels. Useful for
troubleshooting.}

\item{hide_label_if_less_than}{Not yet working.}

\item{font_size}{Font size. Defaults to 8.}

\item{round_digits}{Integer, defaults to 1. Number of decimals in charts and
tables. Currently not working for charts.}

\item{show_n}{Logical, defaults to FALSE. Should number of cases in a
category/group be added to axis labels? NA means TRUE where there is
variation across groups/categories, FALSE otherwise.}

\item{show_p}{Logical, defaults to TRUE. Should percentage symbol (%) be
added to data labels?}

\item{sorting}{Which sorting variable to use? NULL equals data as it is.}

\item{drop_na_y, drop_na_x}{Logical, defaults to TRUE. Whether to drop missing
as a category.}

\item{type}{1 for horizontal percent-bar for each category. 2 for
average and 3 for median, primarily for interval/continuous/binary
variables.}

\item{drop_duplicates}{Logical, defaults to FALSE. If there are duplicate
rows in design_frame, drop duplicates silently?}

\item{remove_empty_y_var}{Logical, defaults to FALSE. As function cannot
handle rows in design_frame with missing y_var, will drop these rows if
FALSE (NOT YET IMPLEMENTED). Otherwise aborts.}
}
\value{
obj, with the barcharts attached in design_frame under column
  "barchart1".
}
\description{
Can add one of two barcharts:
}
\examples{
ex_survey1_obj_new <- add_barchart_mschart(ex_survey1_inf)
}
