% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_caption.R
\name{add_caption}
\alias{add_caption}
\title{Add Caption to Figures or Tables}
\usage{
add_caption(
  obj,
  str_template = "Figure {rownumber}. Barchart of {y}{comp}{x}. {N}.",
  str_pos = " <+> ",
  str_neg = " <-> ",
  str_not = " <0> ",
  show_n = TRUE
)
}
\arguments{
\item{obj}{Infoframe}

\item{str_template}{glue-template string that specifies caption.}

\item{str_pos, str_neg, str_not}{Add comparison based on the est- and pval-variables.}

\item{show_n}{NOt yet implemented.}
}
\value{
Infoframe
}
\description{
Given an infoframe and a string template, create captions.
}
\examples{
ex_survey1_inf_new <- add_caption(ex_survey1_inf)
}
