% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_analysis.R
\name{add_fit}
\alias{add_fit}
\title{Add Variable with Fit Interpretation}
\usage{
add_fit(
  obj,
  cutoff_chisq = 0.05,
  cutoff_CFI = 0.95,
  cutoff_TLI = 0.95,
  cutoff_RMSEA = 0.05,
  cutoff_RMSEA_p05 = 0.05,
  cutoff_SRMR = 0.05,
  drop_raw_fit = TRUE
)
}
\arguments{
\item{obj}{An object as returned from create_infoframe.}

\item{cutoff_chisq}{Cutoff value for the p-value of the chisquare test of 
absolute fit of model to the data,}

\item{cutoff_CFI}{Cutoff value for the Comparative Fit Index of relative fit
of the model to the null model.}

\item{cutoff_TLI}{Cutoff value for the Tucker-Lewis Fit Index of relative fit
of the model to the null model.}

\item{cutoff_RMSEA}{Cutoff value for the Root Mean Square Residual of 
Approximation of relative fit, when penalizing model complexity.
of the model to the null model.}

\item{cutoff_RMSEA_p05}{Cutoff value for the p-value for the RMSEA.}

\item{cutoff_SRMR}{Cutoff-value for the Standardized Root Mean Square 
Residual for the}

\item{drop_raw_fit}{Remove original fit variables and keep interpretation?}
}
\value{
design_frame with fit interpretation added as a column.
}
\description{
Given the design_frame information on fit (run add_analysis first) for
    constructs, add a superficial interpretation of fit given cutoffs. Note
    that this is generally bad scientific practice, but useful for reports 
    and quick analyses. Also note that fit can be improved by dropping poor
    fitting variables, which is not covered by this function.
}
\examples{
#obj <- add_analyses(obj)
#obj <- add_fit(obj)
}
