% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_percentages.R
\name{add_percentages}
\alias{add_percentages}
\title{Add percentages to design frame, with various options, for later processing.}
\usage{
add_percentages(
  df,
  design_frame,
  set = "upper",
  percent = TRUE,
  round_digits = 0
)
}
\arguments{
\item{df}{The data. Extract with myObj[["df"]]}

\item{design_frame}{Design frame as returned from create_infoframe().}

\item{set}{One of 
\itemize{
\item{"top"}{The percentage for the highest category available in the variable.}
\item{"upper"}{The sum of the percentages for the categories above the middle category.}
\item{"mid_upper"}{The sum of the percentages for the categories including and above the middle category.}
\item{"mid_lower"}{The sum of the percentages for the categories including and below the middle category.}
\item{"lower"}{The sum of the percentages for the categories below the middle category.}
\item{"bottom"}{The percentage for the lowest category available in the variable.}
}}

\item{percent}{If TRUE (default), return proportion*100, otherwise proportion.}

\item{round_digits}{Number of decimals.}
}
\description{
Add percentages to design frame, with various options, for later processing.
}
\examples{
# add_percentages(obj)
}
