% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_table.R
\name{add_table}
\alias{add_table}
\title{Add Table}
\usage{
add_table(
  obj,
  prefix_number = FALSE,
  round_digits = 1,
  show_n = FALSE,
  show_p = TRUE,
  sorting = NULL,
  drop_na_y = TRUE,
  drop_na_x = TRUE,
  type = 1,
  remove_empty_y_var = FALSE
)
}
\arguments{
\item{obj}{A list containing df, var_frame and design_frame, as created by
create_infoframe().}

\item{prefix_number}{Whether to add the underlying number index for the
categories as a prefix in the printed chart and table labels. Useful for
troubleshooting.}

\item{round_digits}{Integer, defaults to 1. Number of decimals in charts and
tables. Currently not working for charts.}

\item{show_n}{Logical, defaults to FALSE. Should number of cases in a
category/group be added to axis labels? NA means TRUE where there is
variation across groups/categories, FALSE otherwise.}

\item{show_p}{Logical, defaults to TRUE. Should percentage symbol (%) be
added to data labels?}

\item{sorting}{Which sorting variable to use? NULL equals data as it is.}

\item{drop_na_y, drop_na_x}{Logical, defaults to TRUE. Whether to drop missing
as a category.}

\item{type}{Integer, either 1 or 2. 1 corresponds to a regular percent table
whereas 2 corresponds to average per x_var category.}

\item{remove_empty_y_var}{Logical, defaults to FALSE. As `reporter()` cannot
handle rows in design_frame with missing y_var, will drop these rows if
FALSE (NOT YET IMPLEMENTED). Otherwise aborts.}
}
\value{
obj, with the barcharts attached in design_frame under column
  "barchart1".
}
\description{
Can add one of two tables:
}
\examples{
ex_survey1_inf_new <- add_table(ex_survey1_inf, type=1)
ex_survey1_inf_new <- add_table(ex_survey1_inf_new, type=2)
}
