% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_text.R
\name{add_text}
\alias{add_text}
\title{Adds Text Interpretations to Selected Variables in Design Frame}
\usage{
add_text(
  obj,
  template_xlsx = NULL,
  first_mention = c("high_p", "low_p", "high_n", "small_n", "high_cat", "small_cat",
    "str_first_cat"),
  seed = 1,
  out_col_prefix = "interpret_"
)
}
\arguments{
\item{obj}{An object as generated by `create_infoframe()` or as returned
by the other add_* functions.}

\item{template_xlsx}{Path to translations file. 
#'     Set as NULL for built-in examples, which is also available by calling 
    `data("template_translations", package="surveyreport")`. Save this to 
    disk, modify according to your needs, then read back in using 
    `readxl::read_excel("my_translations.xlsx")`. File must contain the 
    following three columns. 
\itemize{
\item{type}{Types of strings to randomize between. Will also become column 
    names in design_frame.}
\item{regex}{Regular expression: this is used to select filter y_group_label 
    for which the str is to be used.}
\item{str}{A glue-based string template for building sentences.} 
}}

\item{first_mention}{What principle to use for selecting the first category
being mentioned when comparing categories.}

\item{seed}{Temporary seed number for random number generation. If not NULL,
will randomize among options for a given type.}

\item{out_col_prefix}{Prefix to add to output columns. Defaults recommended.}
}
\value{
obj. design_frame will have added a column for each unique `type`,
    named interpret_*.
}
\description{
Based on the inputs, adds semi-random text interpretations for text-based
    report generation.
}
\examples{
ex_data <- readRDS(system.file("extdata", "ex_survey.RDS", 
                   package="surveyreport", mustWork=TRUE))
obj <- create_infoframe(df=ex_data, 
                        y_var=grep("[abdefg]_", names(ex_data), value=TRUE),
                        x_var=grep("^x", names(ex_data), value=TRUE),
                        ordinal_var=grep("^[abdefg]_", names(ex_data), value=TRUE),
                        nominal_var=c("xsex", "xhuman", "h_1", "h_2", "h_3", "h_4"))
obj <- add_text(obj=obj)
obj$design_frame[, grepl("interpret_", names(obj$design_frame))]

}
