% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_valid_infoframe}
\alias{assert_valid_infoframe}
\title{Assert Infoframe As Valid Object}
\usage{
assert_valid_infoframe(obj)
}
\arguments{
\item{obj}{An infoframe object as created by create_infoframe()}
}
\value{
Nothing.
}
\description{
Checks that an object is a valid infoframe object. Currently does not return
    anything.
}
