% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_options}
\alias{check_options}
\title{Mostly an internal function to check whether options provided are valid}
\usage{
check_options(df, df_var, global_default, options)
}
\arguments{
\item{df}{Typically a design frame or a var_frame.}

\item{df_var}{Character string for variable name.}

\item{global_default}{If missing, what to set as default value.}

\item{options}{A type (character()) or a character vector with valid options.}
}
\value{
Data frame, optionally with the missing variable added, set with the
  global_default if it is among the options.
}
\description{
Not much else to say.
}
