% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{col_to_binaries}
\alias{col_to_binaries}
\title{Mutate a (factor, character, integer, etc) column into multiple columns,}
\usage{
col_to_binaries(data, col, var_separator = "___", label_separator = " - ")
}
\arguments{
\item{data}{Data frame or tibble}

\item{col}{Single column. Tidy-select.}

\item{var_separator}{[\code{character(1)>0}]\cr Separator between old variable name and categories.}

\item{label_separator}{[\code{character(1)>0}]\cr Separator between old label name and new label part.}
}
\value{
Original data frame with the extra columns attached.
}
\description{
Easily mutate a single column into multiple columns (~dummies+1),
while retaining variable labels and order of the original factor variable.
}
\examples{
library(dplyr)
library(officer)
ex_survey1 \%>\%
  col_to_binaries(col = b_3, label_separator = "  -  ") \%>\%
  mutate(across(matches("b_3_"), .fns = ~1-.x)) \%>\% # Flip categories
  report_chart_likert(cols = matches("b_3_"), what = "freq") \%>\%
  print(target="test.docx")
}
