% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{colour_picker}
\alias{colour_picker}
\title{Get Colour Palette}
\usage{
colour_picker(type, n_unique, colour_set_ordinal, colour_set_nominal)
}
\arguments{
\item{type}{Character vector of variable types ("ordinal", "nominal", "interval")}

\item{n_unique}{Number of unique values}
}
\value{
A character vector of hex colours.
}
\description{
Give two
}
