% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_likert.R
\name{create_chart_likert}
\alias{create_chart_likert}
\title{Create Likert Chart from Descriptives Data Frame}
\usage{
create_chart_likert(
  data,
  y = "n",
  x = "name",
  group = "value",
  labels = "label",
  label_font_size = 10,
  main_font_size = 8,
  colour_palette = NULL,
  seed = 1
)
}
\arguments{
\item{data}{Data frame or tibble.}

\item{y}{[\code{character(1)}]\cr Name of column in data for .}

\item{x}{[\code{character(1)}]\cr Name of column in data for}

\item{group}{[\code{character(1)}]\cr Name of column in data for}

\item{labels}{[\code{character(1)}]\cr Name of column in data for labels.}

\item{label_font_size}{[\code{integer(1)}]\cr Font size for data labels}

\item{main_font_size}{[\code{integer(1)}]\cr Font size for all other text}

\item{colour_palette}{[\code{character()}]\cr}

\item{seed}{Optional random seed for selection of colours in blender.
Must contain at least the number of unique values (incl. missing) in the data set.}
}
\value{
mschart-object. Can be added to an rdocx, rpptx or rxlsx object.
}
\description{
Create Likert Chart from Descriptives Data Frame
}
\examples{
data(ex_survey1)
dat <- ex_survey1[,c("b_1", "b_2", "b_3")]
dat_freq <- prepare_freq_data(dat)
create_chart_likert(dat_freq)
}
