% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_plot.R
\name{create_forest_plot}
\alias{create_forest_plot}
\title{Creates List with Forest Plots from Data Frame}
\usage{
create_forest_plot(
  data,
  caption_suffix = "",
  y_string_wrap_width = 40,
  reverse_y = TRUE
)
}
\arguments{
\item{data}{Data frame outputed from extract_mplus_reg_coefs()}

\item{caption_suffix}{Optional suffix string}

\item{y_string_wrap_width}{Wrap y_labels if wider than this.}

\item{reverse_y}{Reverse order of y_axis. Defaults to TRUE.}
}
\value{
List with ggplots
}
\description{
Creates List with Forest Plots from Data Frame
}
