% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_infoframe.R
\name{create_infoframe}
\alias{create_infoframe}
\title{Create An Infoframe}
\usage{
create_infoframe(
  df,
  var_group_item_sep = "_",
  varlabel_group_item_sep = " - ",
  y_var = grep("^x", colnames(df), value = TRUE, invert = TRUE),
  x_var = NULL,
  med_var = NULL,
  ordinal_var = NULL,
  nominal_var = NULL,
  interval_var = NULL,
  text_var = NULL,
  add_constructs = TRUE,
  add_x_univariates = FALSE,
  add_y_univariates = FALSE,
  add_m_univariates = FALSE,
  ordinal_if_fewer_than = 5,
  vars_as_labels = FALSE,
  colour_set_ordinal = c("#E8B0B7", "#DE919A", "#D5727D", "#C84957", "#BC3848",
    "#9D2F3C", "#7E2630"),
  colour_set_nominal = c(red = "#C82D49", black = "#363636", beige = "#EDE2D2", blue =
    "#2D8E9F", purple = "#DBD2E0")
)
}
\arguments{
\item{df}{A labelled data frame containing only the data columns one wishes 
to create item groups for.}

\item{var_group_item_sep}{Separator between item group identifier prefix 
and item suffix. Default="_" is currently the only that works well for 
separating. Set "$" if each item is to be considered a group.}

\item{varlabel_group_item_sep}{Same as for `var_group_item_sep`, but for 
separating labels into group and item specific.}

\item{y_var}{character vector with dependent data columns.}

\item{x_var}{character vector with independent data columns.}

\item{med_var}{character vector with mediator data columns.}

\item{ordinal_var}{character vector with data columns having ordinal meaning.}

\item{nominal_var}{character vector with data columns having nominal meaning.}

\item{interval_var}{character vector with data columns having interval meaning.}

\item{text_var}{character vector with data columns having text meaning 
(will be excluded from figures, tables, analyses, but will be reported 
in wordclouds and as-is.}

\item{add_constructs}{Constructs consist of multiple data columns measuring 
essentially the same. Add these as new construct based on group-variable?}

\item{add_y_univariates, add_x_univariates, add_m_univariates}{Whether or not 
to report univariates for dependent/independent/mediator data columns.}

\item{ordinal_if_fewer_than}{???}

\item{vars_as_labels}{Logical, TRUE (default) or FALSE). If TRUE, will use 
data column names as labels if labels are missing.}

\item{colour_set_ordinal}{A character vector with hex-colours used for 
ordinal data columns in charts/plots.}

\item{colour_set_nominal}{A character vector with hex-colours used for 
nominal data columns in charts/plots.}
}
\value{
A list consisting of the original data, a var_frame (codebook for 
    each data column), and a design_frame.
\itemize{
\item{var_frame}{The var_frame (codebook).}
\itemize{
\item{group}{Data column group.}
\item{item}{Original data column.}
\item{group_label}{Group variable label.}
\item{item_label}{Original data column label.}
\item{type}{Data column type: ordinal, nominal, text, interval, 
    singular (only one value)}
}
\item{design_frame}
\itemize{
\item{y_var}
\item{y_label}
\item{y_group}
\item{y_group_label}
\item{x_var}
\item{x_label}
\item{x_group}
\item{x_group_label}
\item{y_type}
\item{x_type}
}
}
}
\description{
Will create a list with the original data, a data column frame (codebook), 
    and a design frame for y-x analyses, tables, figures and reports. 
    Defaults in the design frame can be modified later. This function should
    be run if the original data has been changed.
}
\examples{
ex_data <- readRDS(system.file("extdata", "ex_survey.RDS", 
                   package="surveyreport", mustWork=TRUE))
obj <- create_infoframe(df=ex_data, 
                        y_var=grep("^[abdefg]_", names(ex_data), value=TRUE),
                        x_var=grep("^x", names(ex_data), value=TRUE),
                        ordinal_var=grep("^[abdefg]_", names(ex_data), value=TRUE),
                        nominal_var=c("xsex", "xhuman", "h_1", "h_2", "h_3", "h_4"),
                        add_x_univariates=TRUE,
                        add_y_univariates=TRUE)
obj$var_frame
obj$design_frame
}
