% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report.R
\name{create_report}
\alias{create_report}
\title{Generate Report From Template and Design Frame}
\usage{
create_report(
  obj,
  section_divider1 = NULL,
  section_divider2 = NULL,
  sorting = NULL,
  remove_empty_y_var = FALSE,
  drop_duplicates = TRUE,
  filepath = "tmp.docx",
  docx_template_path = "template/NIFUmal_tom.docx",
  pptx_template_path = "template/NIFUmal_tom.pptx",
  xlsx_template_path = "template/NIFUmal_tom.xlsx",
  template_styles = "template/NIFUmal_stiler.xlsx"
)
}
\arguments{
\item{obj}{A list containing df, var_frame and design_frame, as created by
create_infoframe().}

\item{section_divider1, section_divider2}{Variable in design_frame that
specifies under which section (level2) and subsection (level3) in the
docx-document that the results will be reported.}

\item{sorting}{Sorting order for results within a section2}

\item{remove_empty_y_var}{Logical, defaults to FALSE. As `reporter()` cannot
handle rows in design_frame with missing y_var, will drop these rows if
FALSE (NOT YET IMPLEMENTED). Otherwise aborts.}

\item{drop_duplicates}{Logical, defaults to FALSE. If there are duplicate
rows in design_frame, drop duplicates silently?}

\item{filepath}{Directory to save file(s).}

\item{docx_template_path, pptx_template_path, xlsx_template_path}{File path 
to docx/pptx/xlsx templates.}

\item{template_styles}{Either a data frame or filepath with two columns:
`surveyreport_style` (must be as is), and `template_style` (must fit styles
in the documents, see NIFUmal_stiler.xlsx for all available).}
}
\value{
design_frame
}
\description{
Generates a report from infoframe object. If specifics have not been set per
row in obj$design_frame, will then use global arguments in this function as
defaults. Can optionally be pre-processed with add_percentages(),
add_analyses(), and add_text().
}
\examples{
ex_survey1_inf_new <- add_caption(ex_survey1_inf)
create_report(ex_survey1_inf_new,
  filepath=file.path(tempdir(), tempfile(fileext = ".docx")))
}
