% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_likert.R
\name{cross_n}
\alias{cross_n}
\title{Title}
\usage{
cross_n(data, showNA = "ifany")
}
\arguments{
\item{data}{Data frame.}

\item{showNA}{"ifany", "always", "no"}
}
\value{
data frame.
}
\description{
Title
}
