% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{crosstable_to_apa}
\alias{crosstable_to_apa}
\title{Title}
\usage{
crosstable_to_apa(
  x,
  label_separator = NULL,
  docx_template = NULL,
  caption_style = NULL,
  font_size = 11,
  font_family = "Times New Roman",
  caption_autonum = NULL,
  topcaption = TRUE
)
}
\arguments{
\item{x}{Crosstable}

\item{label_separator}{[\code{character(1)}]\cr If not NULL (default), will split labels into main- and sub-questions and create figure caption.}

\item{docx_template}{[\code{character(1) || officer::read_docx()}]\cr}

\item{caption_style}{[\code{character(1)}]\cr Template style to be used for formatting chart. Defaults to "Normal".}

\item{font_size}{[\code{integer(1)}]\cr Font size for all text.}

\item{font_family}{[\code{character(1)}]\cr Office font family. Defaults to "Arial". See ?officer::fp_text() for options.}

\item{caption_autonum}{Object obtained from \link[officer]{run_autonum}.}

\item{topcaption}{[\code{logical(1)}]\cr Place caption above (TRUE, default) or below table.}
}
\value{
rdocx object, which can be saved with officer:::print.rdocx()
}
\description{
Title
}
\examples{
library(crosstable)
library(dplyr)
library(officer)
crosstable::crosstable(ex_survey1, b_1:b_3, percent_pattern = "{p_col}") \%>\%
  crosstable::pivot_crosstable() \%>\%
  crosstable_to_apa(label_separator=" - ") \%>\%
  print(target = "test.docx")
  file.remove("test.docx")
}
