% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{db_close}
\alias{db_close}
\title{Close the connection to the PostgreSQL database}
\usage{
db_close(drv, con)
}
\arguments{
\item{drv}{database driver}

\item{con}{database connection}
}
\description{
Close the connection to the PostgreSQL database
}
