% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{db_connection}
\alias{db_connection}
\title{Open connection to the PostgreSQL database}
\usage{
db_connection()
}
\value{
A list containing the database driver and connection.
}
\description{
Open connection to the PostgreSQL database
}
