% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_survey1_inf}
\alias{ex_survey1_inf}
\title{Example Infoframe Created from ex_survey1}
\format{
list containing a dataframe (df), a codebook (var_frame) and a design frame (design_frame).

\describe{list}
}
\usage{
ex_survey1_inf
}
\description{
Ready stored infoframe list object made in the example at \code{\link{create_infoframe}}.
}
\keyword{datasets}
