% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex_survey2}
\alias{ex_survey2}
\title{ex_survey2: Anonymized snippet of a real survey.}
\format{
A data frame with 100 rows and 29 variables: 
\describe{
  \item{x_sex}{Gender} 
}
}
\usage{
ex_survey2
}
\description{
A dataset containing real respondents' answers to survey questions. The
separator between items and groups is the underscore _ (i.e. a_1-a_9 => a +
1-9) whereas the separator for labels is the hyphen. The latter corresponds
with the default in SurveyXact. Note that _ separator will not work 
perfectly for distinguishing all variables and groups.
}
\seealso{
ex_survey1
}
\keyword{datasets}
