% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_analyses.R
\name{extract_mplus_error}
\alias{extract_mplus_error}
\title{Extract Errors From MplusAutomation Model Object}
\usage{
extract_mplus_error(mod)
}
\arguments{
\item{mod}{MplusAutomation model object.}
}
\value{
String
}
\description{
First version of this is very crude, simply extracting every error and 
    returning a string.
}
