% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_analysis.R
\name{extract_mplus_model}
\alias{extract_mplus_model}
\title{Extract MplusAutomation Results Into A Single Row}
\usage{
extract_mplus_model(mod, y_var, y_group, x_var, x_group)
}
\arguments{
\item{mod}{mplusAutomation.model object.}

\item{y_var, x_var}{Character vector with dependent/independent variables to
extract.}

\item{y_group, x_group}{String with name of factor.}
}
\value{
A single data frame row with most important information.
}
\description{
Takes a MplusAutomation object and names of variables to extract. If model 
    contains error, will attempt to extract these.
}
