% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_colour_set}
\alias{get_colour_set}
\title{Provide A Colour Set for A Number of Requested Colours}
\usage{
get_colour_set(
  n_colours_needed,
  user_colour_set = NULL,
  seed = 1,
  call = rlang::caller_env()
)
}
\arguments{
\item{n_colours_needed}{Number of colours needed.}

\item{user_colour_set}{User-supplied default palette.}

\item{seed}{Random seed for sampling.}

\item{call}{Caller function if used from inside another function.}
}
\value{
A colour set as character vector.
}
\description{
Possibly using user_colour_set if available. If not sufficient, uses a set
    palette from RColorBrewer.
}
\examples{
get_colour_set(n_colours_needed=4)
}
