% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report.R
\name{get_template_styles}
\alias{get_template_styles}
\title{Generates Template Styles Object From Excel File}
\usage{
get_template_styles(template_styles)
}
\arguments{
\item{template_styles}{Either a data frame or a link to an Excel file that
contains the columns surveyreport_style and template_style.}
}
\value{
A list object.
}
\description{
This is an internal function to check and obtain the template_styles that
    is given as argument to the reporter() function.
}
