% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_cols.R
\name{handpick}
\alias{handpick}
\title{Return character vector of manually picked data columns.}
\usage{
handpick(data, to_clipboard = FALSE)
}
\arguments{
\item{data}{A dataset; tibble or data frame.}

\item{to_clipboard}{Boolean. Defaults to FALSE. If TRUE, overwrites what you already have copied. Use with caution.}
}
\value{
Character vector. Prints to console.
}
\description{
Return character vector of manually picked data columns.
}
\examples{
if(interactive()) handpick(mtcars)
}
