% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_barchart_mschart.R
\name{ms_percentbar}
\alias{ms_percentbar}
\title{Create a MS Chart Horizontal Percent Barplot}
\usage{
ms_percentbar(
  df,
  y_var,
  x_var,
  group_string = "category",
  user_colours = NULL,
  f_size = 8,
  means = FALSE
)
}
\arguments{
\item{df}{Data frame}

\item{y_var, x_var}{Variables for y and x axes.}

\item{group_string}{Category to group on.}

\item{user_colours}{Character vector with hex colours.}

\item{f_size}{Font size, numeric value.}

\item{means}{Logical, default is FALSE. Whether to show means or percent.}
}
\value{
A mschart object for a rdocx.
}
\description{
Mostly an internal function.
}
