% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_likert.R
\name{prepare_data_for_mschart}
\alias{prepare_data_for_mschart}
\title{Helper Function to Prepare Data for create_chart_likert}
\usage{
prepare_data_for_mschart(data)
}
\arguments{
\item{data}{Dataset}
}
\value{
Dataset
}
\description{
Helper Function to Prepare Data for create_chart_likert
}
