% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepare_freq_data}
\alias{prepare_freq_data}
\title{Prepare Frequencies Table for Likert Graphs}
\usage{
prepare_freq_data(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
Data frame.
}
\description{
Prepare Frequencies Table for Likert Graphs
}
\examples{
prepare_freq_data(ex_survey1[, paste0("p_", 1:4)])
}
