% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_likert.R
\name{prepare_freq_for_mschart}
\alias{prepare_freq_for_mschart}
\title{Helper Function to Prepare Frequency Data for create_chart_likert}
\usage{
prepare_freq_for_mschart(
  data,
  showNA = "ifany",
  call = rlang::caller_env(),
  sort_by = NULL,
  desc = FALSE
)
}
\arguments{
\item{data}{Dataset}

\item{showNA}{Whether to show NA in categorical variables (one of c("ifany", "always", "no"), like in table()).}

\item{call}{Error call function, usually not needed.}

\item{sort_by}{String, sort by either "value", "label", ".id" (variable name) or NULL (default).}

\item{desc}{Reverse sorting of sort_by}
}
\value{
Dataset
}
\description{
Helper Function to Prepare Frequency Data for create_chart_likert
}
