% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_analyses.R
\name{prepare_mplus_model}
\alias{prepare_mplus_model}
\title{Prepare Mplus Model Syntax and Run Model}
\usage{
prepare_mplus_model(
  df,
  y_var,
  x_var,
  y_group,
  x_group,
  y_type,
  x_type,
  cluster,
  estimator = c("wlsmv", "mlr", NA),
  path
)
}
\arguments{
\item{df}{Data frame.}

\item{y_var}{String.}

\item{x_var}{String.}

\item{y_group}{String.}

\item{x_group}{String.}

\item{y_type}{String.}

\item{x_type}{String.}

\item{cluster}{String. Optional.}

\item{estimator}{String. One of "wlsmv", "mlr", or NA}

\item{path}{Path to Mplus files. Must be provided.}
}
\value{
Data frame row.
}
\description{
Internal function.
}
