% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_likert.R
\name{prepare_mschart_data}
\alias{prepare_mschart_data}
\title{Helper Function to Prepare Data for create_chart_likert}
\usage{
prepare_mschart_data(
  data,
  cols = tidyselect::everything(),
  by = NULL,
  what = "percent",
  showNA = "ifany",
  call = rlang::caller_env(),
  digits = 1,
  percent_sign = TRUE,
  sort_by = NULL,
  desc = FALSE,
  hide_label_if_below = 1,
  label_separator = NULL
)
}
\arguments{
\item{data}{Dataset}

\item{cols}{Columns to select for reporting. Supports \code{\link[dplyr:dplyr_tidy_select]{tidy-select}}.}

\item{by}{\code{\link[dplyr:dplyr_data_masking]{data-masking}}\cr. Optional column used to break graph by.}

\item{what}{[\code{character(1)}]\cr Whether to compute "percentage" (default) or "frequency". Supports partial matching.}

\item{showNA}{[\code{logical(1)}]\cr Whether to show NA in categorical variables (one of c("ifany", "always", "no"), like in table()).}

\item{call}{Error call function, usually not needed.}

\item{digits}{[\code{integer(1)}]\cr Number of decimal places as integer.}

\item{percent_sign}{[\code{logical(1)}]\cr Whether to include percentage symbol on chart.}

\item{sort_by}{[\code{character(1)}], sort output by "label", size of a category, or sum of categories (character vector). Defaults to none (NULL).}

\item{desc}{[\code{logical(1)}]\cr Reverse sorting of sort_by. Defaults to ascending order (FALSE).}

\item{hide_label_if_below}{[\code{numeric(1)}] Whether to hide label if below this value.}

\item{label_separator}{[\code{character(1)}]\cr Split pattern.}
}
\value{
Dataset
}
\description{
Helper Function to Prepare Data for create_chart_likert
}
