% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_likert.R
\name{prepare_perc_for_mschart}
\alias{prepare_perc_for_mschart}
\title{Helper Function to Prepare Data for create_chart_likert}
\usage{
prepare_perc_for_mschart(
  data,
  showNA = "ifany",
  call = rlang::caller_env(),
  digits = 1,
  percent_sign = TRUE,
  sort_by = NULL,
  desc = FALSE,
  hide_label_if_below = 1,
  label_separator = NULL
)
}
\arguments{
\item{data}{Dataset}

\item{showNA}{Whether to show NA in categorical variables (one of c("ifany", "always", "no"), like in table()).}

\item{call}{Error call function, usually not needed.}

\item{digits}{Number of decimal places as integer.}

\item{percent_sign}{Logical, whether to include percentage symbol on chart.}

\item{sort_by}{[\code{character(1)}], sort output by "label", size of a category, or sum of categories (character vector). Defaults to none (NULL).}

\item{desc}{Reverse sorting of sort_by}

\item{hide_label_if_below}{[\code{numeric(1)}] Whether to hide label if below this value.}

\item{label_separator}{Split pattern.}
}
\value{
Dataset
}
\description{
Helper Function to Prepare Data for create_chart_likert
}
