% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_report.R
\name{print2.rdocx}
\alias{print2.rdocx}
\title{Prints rdocx and fixes issues with mschart}
\usage{
print2.rdocx(x, target = file.path(getwd(), "tmp.docx"))
}
\arguments{
\item{x}{rdocx object}

\item{target}{target file path}
}
\value{
target file path
}
\description{
mschart package has a small bug which this function rectifies.
}
