% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_likert.R
\name{report_chart_likert}
\alias{report_chart_likert}
\title{Create Word Report with Univariates for Categorical Columns Sharing Same Categories}
\usage{
report_chart_likert(
  data,
  cols,
  docx_template = NULL,
  label_font_size = 8,
  colour_palette = NULL,
  chart_formatting = NULL,
  height_per_col = 0.3,
  height_fixed = 1,
  main_font_size = 8,
  seed = 1
)
}
\arguments{
\item{data}{Data frame or tibble.}

\item{cols}{<tidy-select> Columns to select for reporting.}

\item{docx_template}{[\code{character(1) || officer::read_docx()}]\cr
Either a filepath to a template file, or a rdocx-object.}

\item{label_font_size}{[\code{integer(1)}]\cr Font size for data labels}

\item{colour_palette}{[\code{character()}]\cr
Must contain at least the number of unique values (incl. missing) in the data set.}

\item{chart_formatting}{[\code{integer(1)}]\cr
Which template style to be used for formatting chart?}

\item{height_per_col}{[\code{numeric(1)>0}]\cr Height in cm per chart entry.}

\item{height_fixed}{[\code{numeric(1)>0}]\cr Fixed height in cm.}

\item{main_font_size}{[\code{integer(1)}]\cr Font size for all other text}

\item{seed}{Optional random seed for selection of colours in blender.}
}
\description{
Create Word Report with Univariates for Categorical Columns Sharing Same Categories
}
