% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter.R
\name{reporter}
\alias{reporter}
\title{Generate Report From Template and Design Frame}
\usage{
reporter(
  obj,
  section_divider1 = NULL,
  section_divider2 = NULL,
  sorting = NULL,
  y_colour_set = NULL,
  drop_na_y = TRUE,
  drop_na_x = TRUE,
  prefix_number = FALSE,
  remove_empty_y_var = FALSE,
  drop_duplicates = TRUE,
  hide_label_if_less_than = 0,
  font_size = 8,
  round_digits = 1,
  show_n_chart = c("caption", "axis", "auto", "none"),
  show_n_table = c("caption", "axis", "auto", "none"),
  show_p_table = c("caption", "axis", "auto", "none"),
  create_barplot1 = TRUE,
  create_barplot2 = FALSE,
  create_table1 = TRUE,
  create_table2 = FALSE,
  docx = TRUE,
  pptx = FALSE,
  xlsx = FALSE,
  str_pos = " <+> ",
  str_neg = " <-> ",
  str_not = " <0> ",
  str_blank = 0,
  str_table_tag = "Tabell ",
  str_figure_tag = "Figur ",
  path = getwd(),
  file_prefix = "uni",
  docx_template_path = "template/NIFUmal_tom.docx",
  pptx_template_path = "template/NIFUmal_tom.pptx",
  xlsx_template_path = "template/NIFUmal_tom.xlsx",
  template_styles = "template/NIFUmal_stiler.xlsx"
)
}
\arguments{
\item{obj}{A list containing df, var_frame and design_frame, as created by
create_infoframe().}

\item{section_divider1, section_divider2}{Variable in design_frame that
specifies under which section (level2) and subsection (level3) in the
docx-document that the results will be reported.}

\item{sorting}{Sorting order for results within a section2}

\item{y_colour_set}{Character vector with hex colours for the set of y_var.}

\item{drop_na_y, drop_na_x}{Logical, defaults to TRUE. Whether to drop missing
as a category.}

\item{prefix_number}{Whether to add the underlying number index for the
categories as a prefix in the printed chart and table labels. Useful for
troubleshooting.}

\item{remove_empty_y_var}{Logical, defaults to FALSE. As `reporter()` cannot
handle rows in design_frame with missing y_var, will drop these rows if
FALSE (NOT YET IMPLEMENTED). Otherwise aborts.}

\item{drop_duplicates}{Logical, defaults to FALSE. If there are duplicate
rows in design_frame, drop duplicates silently?}

\item{hide_label_if_less_than}{Not yet working.}

\item{font_size}{Font size. Defaults to 8.}

\item{round_digits}{Integer, defaults to 1. Number of decimals in charts and
tables. Currently not working for charts.}

\item{show_n_chart}{Logical, defaults to "caption". Should number of cases in
a category/group be added? One of "axis" (part of the chart axis text),
"caption" (summarized and added to the caption), "none" (off) or "auto"
(caption if no variation, otherwise in axis).}

\item{show_n_table}{Same as show_n_chart, but for the tables.}

\item{show_p_table}{Show percentage symbol (%) in table header ("axis") or in
the caption ("caption").}

\item{create_barplot1}{Logical, defaults to TRUE. Percents in each category.}

\item{create_barplot2}{Logical, defaults to TRUE. Percents in each category.}

\item{create_table1}{Logical, defaults to FALSE. Percents in each category.}

\item{create_table2}{Logical, defaults to TRUE. Means for each category.}

\item{docx}{Logical, defaults to TRUE. Whether to make Word document.}

\item{pptx}{Logical, defaults to FALSE Whether to make PowerPoint document.}

\item{xlsx}{Logical, defaults to FALSE Whether to make Excel document.}

\item{str_pos, str_neg, str_not, str_blank}{String specifying how relations
between y and x are to be presented in captions for positive, negative, no
relations, and for univariates (str_blank).}

\item{str_table_tag, str_figure_tag}{String specifying what "Table " and
"Figure " should be displayed as in captions.}

\item{path}{Directory to save file(s).}

\item{file_prefix}{Filename prefix (before file extension).}

\item{docx_template_path, pptx_template_path, xlsx_template_path}{File path 
to docx/pptx/xlsx templates.}

\item{template_styles}{Either a data frame or filepath with two columns:
`surveyreport_style` (must be as is), and `template_style` (must fit styles
in the documents, see NIFUmal_stiler.xlsx for all available).}
}
\value{
design_frame
}
\description{
Generates a report from infoframe object. If specifics have not been set per
row in obj$design_frame, will then use global arguments in this function as
defaults. Can optionally be pre-processed with add_percentages(),
add_analyses(), and add_text().
}
