% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporter.R
\name{reverse_unlabeling}
\alias{reverse_unlabeling}
\title{Not sure what this is for yet}
\usage{
reverse_unlabeling(var)
}
\arguments{
\item{var}{Vector}
}
\value{
A factor.
}
\description{
Some function for reverse unlabelling?
}
