% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_analysis.R
\name{run_single_analysis}
\alias{run_single_analysis}
\title{Run Single Analyses in Design Frame}
\usage{
run_single_analysis(
  df,
  adjust_dummy01,
  y_var,
  y_group,
  y_type = c("interval", "nominal", "ordinal"),
  x_var,
  x_group,
  x_type = c("interval", "nominal", "ordinal"),
  x_ref_cat,
  cluster,
  engine = c("mplus", "lavaan", "lm"),
  estimator = c("wlsmv", "mlr", NA),
  ...
)
}
\arguments{
\item{df}{Data frame where y_var, x_var, and optionally cluster can be found.}

\item{adjust_dummy01}{Whether to fix dummies with lowest category above 0.}

\item{y_var}{Character vector, must be provided and refer to columns in df.}

\item{y_group}{String, optional.}

\item{y_type}{Character vector, optional. Must be a single value or the same
length as y_var.}

\item{x_var}{String, optional.}

\item{x_group}{String, optional.}

\item{x_type}{Character vector, optional. Must be a single value or the same
length as x_var.}

\item{x_ref_cat}{Optional string indicating x_var category used as reference,
if x_var type is nominal or ordinal.}

\item{cluster}{Optional global default.}

\item{engine}{String, one of "wlsmv", "mlr" or NA.}

\item{estimator}{One of "wlsmv", "mlr", or NA. WLSMV is Weighted Least
Squares with Means and Variance Adjustment, suitable for non-normally
distributed dependent variables. Robust Maximum Likelihood (mlr) is 
appropriate if the missing data is missing at random, e.g. can be 
explained by the available variables in the model.}

\item{...}{Other variables accidentally sent to function.}
}
\value{
Single row data frame.
}
\description{
Given a row in a design frame, runs it.
}
