% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_likert.R
\name{sorter}
\alias{sorter}
\title{Internal: Finalizes dataset. Helper function.}
\usage{
sorter(
  data,
  sort_by = NULL,
  desc = FALSE,
  hide_label_if_below = 1,
  label_separator = NULL
)
}
\arguments{
\item{data}{Dataset}

\item{sort_by}{[\code{character(1)}], sort output by "label", size of a category, or sum of categories (character vector). Defaults to none (NULL).}

\item{desc}{Logical, defaults to ascending order (FALSE).}

\item{hide_label_if_below}{[\code{numeric(1)}] Whether to hide label if below this value.}

\item{label_separator}{Split pattern.}
}
\value{
dataset
}
\description{
Internal: Finalizes dataset. Helper function.
}
