% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset_vector}
\alias{subset_vector}
\title{Given Ordered Integer Vector, Return Requested Set.}
\usage{
subset_vector(
  vec,
  set = c("top", "upper", "mid_upper", "lower", "mid_lower", "bottom")
)
}
\arguments{
\item{vec}{A vector of any type.}

\item{set}{A character string, one of c("top", "upper", "mid_upper", "lower",
"mid_lower", "bottom")}
}
\value{
Selected set of vector.
}
\description{
Useful for identifying which categories are to be collected.
}
\examples{
subset_vector(vec=1:7, set="mid_lower")
}
