% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observe_multiple_comb.R
\name{t_test_multiple_comb}
\alias{t_test_multiple_comb}
\title{Extention to the infer package for multiple testing}
\usage{
t_test_multiple_comb(
  x,
  formula = NULL,
  response = NULL,
  explanatory = NULL,
  order = NULL,
  alternative = "two-sided",
  mu = 0,
  conf_int = TRUE,
  conf_level = 0.95
)
}
\arguments{
\item{x}{A data frame that can be coerced into a \link[tibble:tibble]{tibble}.}

\item{formula}{A formula with the response variable on the left and the
explanatory on the right.}

\item{response}{The variable name in \code{x} that will serve as the response.
This is alternative to using the \code{formula} argument.}

\item{explanatory}{The variable name in \code{x} that will serve as the
explanatory variable.}

\item{order}{A string vector of specifying the order in which the levels of
the explanatory variable should be ordered for subtraction, where \code{order = c("first", "second")} means \code{("first" - "second")}.}

\item{alternative}{Character string giving the direction of the alternative
hypothesis. Options are \code{"two-sided"} (default), \code{"greater"}, or \code{"less"}.}

\item{mu}{A numeric value giving the hypothesized null mean value for a one
sample test and the hypothesized difference for a two sample test.}

\item{conf_int}{A logical value for whether to include the confidence
interval or not. \code{TRUE} by default.}

\item{conf_level}{A numeric value between 0 and 1. Default value is 0.95.}
}
\value{
A tibble containing for each response-explanatory row combination
columns summarizing the test statistic as obtained from broom::tidy().
}
\description{
Extention to the infer package for multiple testing
}
\examples{
library(dplyr)
x <- t_test_multiple_comb(x = mtcars,
                     response = matches("mpg|disp|drat|hp|qsec"),
                     explanatory = matches("vs|am"))
}
