% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{tab_cat_list_to_plots}
\alias{tab_cat_list_to_plots}
\title{Turn a list of tabulated data frames into plots to be saved.}
\usage{
tab_cat_list_to_plots(
  l,
  y_string_width = 40,
  y_reverse = TRUE,
  output_dir = getwd(),
  output_file_prefix = "",
  output_ext = "png",
  plot_width = 33.8,
  plot_height = 19,
  plot_scale = 1,
  plot_units = "cm",
  base_font_size = 12
)
}
\arguments{
\item{l}{List of data frames from tab_cat_prop()}

\item{y_string_width}{y_var strng wrap width}

\item{y_reverse}{Reverse order of y_var? Defaults to true.}

\item{output_dir}{String.}

\item{output_file_prefix}{String}

\item{output_ext}{Format}

\item{plot_width}{Plot width}

\item{plot_height}{Plot height}

\item{plot_scale}{Plot scale}

\item{plot_units}{units. defaults to cm}

\item{base_font_size}{Base font size. Defaults to 12}
}
\value{
List of plots
}
\description{
Turn a list of tabulated data frames into plots to be saved.
}
