% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{tab_cat_prop_to_barplot}
\alias{tab_cat_prop_to_barplot}
\title{Create barplot}
\usage{
tab_cat_prop_to_barplot(
  data,
  y_string_wrap_width = 40,
  reverse_y = FALSE,
  base_font_size = 12,
  percent = FALSE,
  decimals = 3
)
}
\arguments{
\item{data}{data.frame}

\item{y_string_wrap_width}{Wrap y-labels if longer than this. Default=40.}

\item{reverse_y}{Reverse order of y-labels (defaults to FALSE).}

\item{base_font_size}{Base font size. Default=12.}

\item{percent}{Logical, percentage vs proportion (default).}

\item{decimals}{Number of decimals for labels.}
}
\value{
ggplot2-object
}
\description{
Create barplot
}
\examples{
tab_cat_prop_to_barplot(
tab_cat_prop(ex_survey1[,paste0("b_", 1:3)]))
}
