% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{tabyl_lab}
\alias{tabyl_lab}
\title{Fix to the janitor::tabyl function to make it work seamlessly with labelled variables.}
\usage{
tabyl_lab(dat, ..., show_na = TRUE, show_missing_levels = TRUE)
}
\arguments{
\item{dat}{a data.frame containing the variables you wish to count. Or, a vector you want to tabulate.}

\item{...}{the arguments to tabyl (here just for the sake of documentation compliance, as all arguments are listed with the vector- and data.frame-specific methods)}

\item{show_na}{should counts of NA values be displayed? In a one-way tabyl, the presence of NA values triggers an additional column showing valid percentages(calculated excluding NA values).}

\item{show_missing_levels}{should counts of missing levels of factors be displayed? These will be rows and/or columns of zeroes. Useful for keeping consistent output dimensions even when certain factor levels may not be present in the data.}
}
\value{
Returns a data.frame with frequencies and percentages of the tabulated variable(s). A 3-way tabulation returns a list of data.frames.
}
\description{
Fix to the janitor::tabyl function to make it work seamlessly with labelled variables.
}
\examples{
tabyl_lab(mtcars, cyl)
}
