% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{tabyl_multi}
\alias{tabyl_multi}
\title{Generate multiple 1-way frequency tables that share the same response scale.}
\usage{
tabyl_multi(dat, show_na = TRUE, show_missing_levels = TRUE, var_label = NULL)
}
\arguments{
\item{dat}{a data.frame containing the variables you wish to count. Or, a vector you want to tabulate.}

\item{show_na}{the arguments to tabyl (here just for the sake of documentation compliance, as all arguments are listed with the vector- and data.frame-specific methods)}

\item{show_missing_levels}{should counts of NA values be displayed? In a one-way tabyl, the presence of NA values triggers an additional column showing valid percentages(calculated excluding NA values).}

\item{var_label}{Defaults to both variable name and variable label, alternatively, TRUE for only label, FALSE for only name.}
}
\value{
Returns a data.frame with frequencies and percentages of the tabulated variable(s).
}
\description{
Generate multiple 1-way frequency tables that share the same response scale.
}
\examples{
tabyl_multi(ex_survey1[,paste0("a_", 1:9)])
}
