% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_likert.R
\name{create_chart_likert}
\alias{create_chart_likert}
\title{Create Likert Chart from Descriptives Data Frame}
\usage{
create_chart_likert(
  data,
  y = "value",
  x = "label",
  group = "variable",
  labels = "data_label",
  label_font_size = 10,
  main_font_size = 8,
  font_family = "Calibri",
  colour_palette = NULL,
  colour_na = "gray90",
  colour_2nd_binary_cat = "#ffffff",
  vertical = FALSE,
  seed = 1
)
}
\arguments{
\item{data}{Data frame or tibble.}

\item{y}{[\code{character(1)}]\cr Name of column in data for .}

\item{x}{[\code{character(1)}]\cr Name of column in data for}

\item{group}{[\code{character(1)}]\cr Name of column in data for}

\item{labels}{[\code{character(1)}]\cr Name of column in data for labels.}

\item{label_font_size}{[\code{integer(1)}]\cr Font size for data labels}

\item{main_font_size}{[\code{integer(1)}]\cr Font size for all other text
Must contain at least the number of unique values (excl. missing) in the data set.}

\item{font_family}{Word font family. See officer::fp_text}

\item{colour_palette}{[\code{character()}]\cr
Must contain at least the number of unique values (incl. missing) in the data set.}

\item{colour_na}{[\code{character(1)}]\cr Colour as a single string.}

\item{colour_2nd_binary_cat}{[\code{character(1)}]\cr Colour for second category in binary variables. Often useful to hide this.}

\item{vertical}{Logical. If FALSE (default), then horizontal.}

\item{seed}{Optional random seed for selection of colours in blender.}
}
\value{
mschart-object. Can be added to an rdocx, rpptx or rxlsx object.
}
\description{
Create Likert Chart from Descriptives Data Frame
}
\examples{
#create_chart_likert(prepare_data_for_mschart(ex_survey1[paste0("b_", 1:3)]))
}
