% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan_mi.R
\name{lavaan_mi}
\alias{lavaan_mi}
\title{Simply Run Measurement Invariance Model in lavaan}
\usage{
lavaan_mi(data, cols, group, ...)
}
\arguments{
\item{data}{Dataset}

\item{cols}{Columns as character vector}

\item{group}{Groupin variable as string}

\item{...}{Not in use}
}
\value{
Tibble
}
\description{
Simply Run Measurement Invariance Model in lavaan
}
