% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_likert.R
\name{prepare_data_for_mschart}
\alias{prepare_data_for_mschart}
\title{Helper Function to Prepare Data for create_chart_likert}
\usage{
prepare_data_for_mschart(
  data,
  showNA = "ifany",
  call = rlang::caller_env(),
  digits = 1,
  percent = TRUE,
  sort_col = NULL,
  desc = FALSE
)
}
\arguments{
\item{data}{Dataset}

\item{showNA}{Whether to show NA in categorical variables (one of c("ifany", "always", "no"), like in table()).}

\item{call}{Error call function, usually not needed.}

\item{digits}{Number of decimal places as integer.}

\item{percent}{Logical, whether to include percentage symbol on chart.}

\item{sort_col}{String, sort by either "value", "label", ".id" (variable name) or NULL (default).}

\item{desc}{Reverse sorting of sort_col}
}
\value{
Dataset
}
\description{
Helper Function to Prepare Data for create_chart_likert
}
