% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_likert.R
\name{report_chart_likert}
\alias{report_chart_likert}
\title{Create Word Report with Univariates for Categorical Columns Sharing Same Categories}
\usage{
report_chart_likert(
  data,
  cols = everything(),
  docx_template = NULL,
  label_font_size = 8,
  colour_palette = NULL,
  colour_na = "gray90",
  chart_formatting = NULL,
  height_per_col = 0.3,
  height_fixed = 1,
  main_font_size = 8,
  seed = 1
)
}
\arguments{
\item{data}{Data frame or tibble.}

\item{cols}{<tidy-select> Columns to select for reporting.}

\item{docx_template}{[\code{character(1) || officer::read_docx()}]\cr
Either a filepath to a template file, or a rdocx-object.}

\item{label_font_size}{[\code{integer(1)}]\cr Font size for data labels}

\item{colour_palette}{[\code{character()}]\cr
Must contain at least the number of unique values (incl. missing) in the data set.}

\item{colour_na}{[\code{character(1)}]\cr Colour as a single string.}

\item{chart_formatting}{[\code{integer(1)}]\cr
Which template style to be used for formatting chart?}

\item{height_per_col}{[\code{numeric(1)>0}]\cr Height in cm per chart entry.}

\item{height_fixed}{[\code{numeric(1)>0}]\cr Fixed height in cm.}

\item{main_font_size}{[\code{integer(1)}]\cr Font size for all other text}

\item{seed}{Optional random seed for selection of colours in blender.}
}
\value{
rdocx object, which can be saved with officer:::print.rdocx()
}
\description{
Create Word Report with Univariates for Categorical Columns Sharing Same Categories
}
\examples{
library(dplyr) # For piping
library(officer) # To save the rdocx object to disk
ex_survey1 \%>\%
  report_chart_likert(cols = a_1:a_9) \%>\%
  print(target = "test_docx_a19.docx")
file.remove("test_docx_a19.docx")


  docx_template <-
      system.file("template","NIFUmal_tom.docx",
                   package = "surveyreport", mustWork = TRUE)
  colour_palette <-
    readxl::read_excel(system.file("template", "NIFUmal_stiler.xlsx",
                                   package = "surveyreport", mustWork = TRUE),
                       sheet = "NIFUblue") \%>\%
    dplyr::pull(hex)
chart_format <-
 system.file("template", "NIFUmal_stiler.xlsx",
             package = "surveyreport", mustWork = TRUE) \%>\%
 readxl::read_excel(., sheet = 1) \%>\%
 dplyr::filter(surveyreport_style == "figure") \%>\%
 dplyr::pull(template_style)

 test_docx_b13 <-
   ex_survey1 \%>\%
   report_chart_likert(cols = b_1:b_3,
                       docx_template = docx_template,
                       colour_palette = colour_palette,
                       chart_formatting = chart_format,
                       height_per_col = .3,
                       height_fixed = 1)
print(test_docx_b13, target = "test_docx_b13.docx")
file.remove("test_docx_b13.docx")
}
