% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_likert.R
\name{report_chart_likert}
\alias{report_chart_likert}
\title{Create Word Report with Univariates for Categorical Columns Sharing Same Categories}
\usage{
report_chart_likert(
  data,
  cols = everything(),
  by = NULL,
  showNA = "ifany",
  docx_template = NULL,
  label_font_size = 8,
  main_font_size = 9,
  font_family = "Calibri",
  colour_palette = NULL,
  colour_na = "gray90",
  colour_2nd_binary_cat = "#ffffff",
  height_per_col = 0.3,
  height_fixed = 1,
  chart_formatting = NULL,
  caption_style = NULL,
  what = "percent",
  digits = 1,
  percent_sign = TRUE,
  sort_by = NULL,
  vertical = FALSE,
  desc = FALSE,
  hide_label_if_below = 1,
  label_separator = NULL,
  caption_autonum = NULL,
  seed = 1
)
}
\arguments{
\item{data}{Data frame or tibble.}

\item{cols}{Columns to select for reporting. Supports \code{\link[dplyr:dplyr_tidy_select]{tidy-select}}.}

\item{by}{\code{\link[dplyr:dplyr_data_masking]{data-masking}}\cr. Optional column used to break graph by.}

\item{showNA}{[\code{logical(1)>0}]\cr Whether to show NA in categorical variables (one of c("ifany", "always", "no"), like in table()).}

\item{docx_template}{[\code{character(1) || officer::read_docx()}]\cr
Either a filepath to a template file, or a rdocx-object from \link[officer]{read_docx}.}

\item{label_font_size}{[\code{integer(1)}]\cr Font size for data labels}

\item{main_font_size}{[\code{integer(1)}]\cr Font size for all other text}

\item{font_family}{[\code{character(1)}]\cr Office font family. Defaults to "Arial". See ?officer::fp_text() for options.}

\item{colour_palette}{[\code{character()}]\cr
Must contain at least the number of unique values (incl. missing) in the data set.}

\item{colour_na}{[\code{character(1)}]\cr Colour as a single string, for NA-values.}

\item{colour_2nd_binary_cat}{[\code{character(1)}]\cr Colour for second category in binary variables. Often useful to hide this.}

\item{height_per_col}{[\code{numeric(1)>0}]\cr Height in cm per chart entry.}

\item{height_fixed}{[\code{numeric(1)>0}]\cr Fixed height in cm.}

\item{chart_formatting}{[\code{character(1)}]\cr Template style to be used for formatting chart}

\item{caption_style}{[\code{character(1)}]\cr Template style to be used for formatting chart. Defaults to "Normal".}

\item{what}{[\code{character(1)}] Either "percent" or "frequency". Supports partial matching.}

\item{digits}{[\code{integer(1)}]\cr Number of decimal places as integer.}

\item{percent_sign}{[\code{logical(1)>0}]\cr Whether to include percentage symbol on chart.}

\item{sort_by}{[\code{character(1)}], sort output by "label", size of a category, or sum of categories (character vector). Defaults to none (NULL).}

\item{vertical}{[\code{logical(1)>0}]\cr If FALSE (default), then horizontal.}

\item{desc}{[\code{logical(1)>0}]\cr Sort in descending order?}

\item{hide_label_if_below}{[\code{numeric(1)}]\cr Whether to hide label if below this value.}

\item{label_separator}{[\code{character(1)}]\cr If not NULL (default), will split labels into main- and sub-questions and create figure caption.}

\item{caption_autonum}{Object obtained from \link[officer]{run_autonum}.}

\item{seed}{[\code{numeric(1)>0}]\cr Optional random seed for selection of colours in blender.}
}
\value{
rdocx object, which can be saved with print() after loading the officer-package
}
\description{
Create Word Report with Univariates for Categorical Columns Sharing Same Categories
}
\examples{
library(dplyr) # For piping
library(officer) # To save the rdocx object to disk
ex_survey1 \%>\%
  report_chart_likert(cols = a_1:a_9) #\%>\%
 # print(target = "test_docx_a19.docx")
#file.remove("test_docx_a19.docx")


  docx_template <-
      system.file("template","NIFUmal_tom.docx",
                   package = "surveyreport", mustWork = TRUE)
  colour_palette <-
    readxl::read_excel(system.file("template", "NIFUmal_stiler.xlsx",
                                   package = "surveyreport", mustWork = TRUE),
                       sheet = "NIFUblue") \%>\%
    dplyr::pull(hex)
chart_format <-
 system.file("template", "NIFUmal_stiler.xlsx",
             package = "surveyreport", mustWork = TRUE) \%>\%
 readxl::read_excel(., sheet = 1) \%>\%
 dplyr::filter(surveyreport_style == "figure") \%>\%
 dplyr::pull(template_style)

 test_docx_b13 <-
   ex_survey1 \%>\%
   report_chart_likert(cols = b_1:b_3,
                       docx_template = docx_template,
                       colour_palette = colour_palette,
                       chart_formatting = chart_format,
                       height_per_col = .3,
                       height_fixed = 1)
#print(test_docx_b13, target = "test_docx_b13.docx")
#file.remove("test_docx_b13.docx")
}
